+++ shared/types/export/index.ts
@@ -1,3 +1,4 @@
 export * from "./agreements";
 export * from "./contributions";
 export * from "./global";
+export * from "./prequalified";