+++ shared/types/export/prequalified.ts
@@ -1,0 +1,11 @@
+import { DocumentElasticWithSource } from "./global";
+
+type Prequalified = {
+  source: "prequalified";
+  variants: string[];
+};
+
+export type PrequalifiedElasticDocument = Omit<
+  DocumentElasticWithSource<Prequalified>,
+  "slug"
+>;