+++ targets/frontend/src/pages/prequalified/index.tsx
@@ -1,0 +1,12 @@
+import { PrequalifiedList } from "src/modules/prequalified";
+import { Layout } from "src/components/layout/auth.layout";
+
+export function PrequalifiedPage() {
+  return (
+    <Layout title="Liste des requêtes Préqualifiés">
+      <PrequalifiedList />
+    </Layout>
+  );
+}
+
+export default PrequalifiedPage;