+++ shared/elasticsearch/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.39.7](https://github.com/SocialGouv/cdtn-admin/compare/v2.39.6...v2.39.7) (2024-03-28)
+
+**Note:** Version bump only for package @socialgouv/cdtn-elasticsearch
+
+## [2.39.6](https://github.com/SocialGouv/cdtn-admin/compare/v2.39.5...v2.39.6) (2024-03-27)
+
+**Note:** Version bump only for package @socialgouv/cdtn-elasticsearch
+
+## [2.39.5](https://github.com/SocialGouv/cdtn-admin/compare/v2.39.4...v2.39.5) (2024-03-26)
+
+**Note:** Version bump only for package @socialgouv/cdtn-elasticsearch
+
+## [2.39.4](https://github.com/SocialGouv/cdtn-admin/compare/v2.39.3...v2.39.4) (2024-03-26)
+
+**Note:** Version bump only for package @socialgouv/cdtn-elasticsearch
+
 ## [2.39.3](https://github.com/SocialGouv/cdtn-admin/compare/v2.39.2...v2.39.3) (2024-03-25)
 
 **Note:** Version bump only for package @socialgouv/cdtn-elasticsearch